requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
bidict==0.23.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
Django==4.0
easyocr==1.7.1
fastapi==0.110.2
filelock==3.14.0
fsspec==2024.3.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.1
intel-openmp==2021.4.0
Jinja2==3.1.3
JPype1==1.5.0
konlpy==0.6.0
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
lxml==5.2.1
marisa-trie==1.1.0
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
preshed==3.0.9
priority==2.0.0
pyclipper==1.3.0.post5
pydantic==2.7.1
pydantic_core==2.18.2
python-bidi==0.4.2
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.23.2
scipy==1.13.0
setuptools==69.5.1
shapely==2.0.4
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
thinc==8.2.3
tifffile==2024.4.24
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.2
weasel==0.3.4
websockets==12.0
wsproto==1.2.0
